+++ app/libraries/Jentil.php
@@ -36,9 +36,7 @@ final class Jentil extends AbstractTheme
 
     protected function getUtilities(): Utilities
     {
-        $this->utilities = $this->utilities ?: new Utilities($this);
-
-        return $this->utilities;
+        return $this->utilities = $this->utilities ?: new Utilities($this);
     }
 
     /**
@@ -58,9 +56,7 @@ final class Jentil extends AbstractTheme
      */
     protected function getMeta(): array
     {
-        $this->meta = $this->meta ?: $this->meta();
-
-        return $this->meta;
+        return $this->meta = $this->meta ?: $this->meta();
     }
 
     /**
@@ -70,10 +66,9 @@ final class Jentil extends AbstractTheme
     {
         $rel_dir = $this->getUtilities()->fileSystem->relativeDir();
 
-        return (
+        return
             ('package' === $mode && $rel_dir) ||
-                ('theme' === $mode && !$rel_dir)
-        );
+            ('theme' === $mode && !$rel_dir);
     }
 
     private function setUpMisc()
@@ -124,6 +119,9 @@ final class Jentil extends AbstractTheme
 
     private function setUpScripts()
     {
+        $this->setups['Scripts\jQuery'] = new Setups\Scripts\jQuery($this);
+        $this->setups['Scripts\jQueryMigrate'] =
+            new Setups\Scripts\jQueryMigrate($this);
         $this->setups['Scripts\FontAwesome'] =
             new Setups\Scripts\FontAwesome($this);
         $this->setups['Scripts\FontAwesomeShim'] =