+++ assets/scripts/core.ts
@@ -1,10 +1,23 @@
 /// <reference path='./global.d.ts' />
 
-(($: JQueryStatic): void => {
-    'use strict'
+namespace Jentil
+{
+    export class App
+    {
+        public constructor(private readonly _j: JQueryStatic)
+        {
+        }
 
-    /**
-     * Add has-js class to `<body>` tag
-     */
-    $('body').removeClass('no-js').addClass('has-js')
-})(jQuery)
+        public run(): void
+        {
+            this.addBodyClasses()
+        }
+
+        private addBodyClasses(): void
+        {
+            this._j('body').removeClass('no-js').addClass('has-js')
+        }
+    }
+}
+
+new Jentil.App(jQuery).run()