+++ assets/scripts/global.d.ts
@@ -8,11 +8,14 @@ declare const jentilRelatedPostsHeadingModIds: string[]
 
 declare const jentilPageLayoutModIds: string[]
 
-declare const jentilMenuL10n: {
-    submenu: string
-}
+declare const jentilMenuL10n: JentilMenuL10n
 
 declare const wp: WP
+
+interface JentilMenuL10n
+{
+    submenu: string
+}
 
 interface WP
 {